+++ projects/client-api-react/src/bindings.js
@@ -42,7 +42,7 @@ const bindingsTable = {
     gravity:				[ PropTypes.oneOf(Array.from({ length: 100 }, (x, i) => i + 1)), 'defaultGravity',  undefined,  undefined],
     scalingRatio:			[ PropTypes.oneOf(Array.from({ length: 100 }, (x, i) => i + 1)), 'defaultScalingRatio', undefined,  undefined],
 
-    //togglePanel:            [ PropTypes.array,      undefined,                   undefined,      'togglePanel'],
+    setTogglePanel:            [ PropTypes.array,       undefined,                   undefined,      'togglePanel'],
     
     ticks:                  [PropTypes.number,       undefined,                   undefined,      'tickClustering'],